+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ### Features / Enhancements
 
 - Added poster for video (#120)
-- Update current index on data series decrease (#121)
+- Updated current index on data series decrease (#121)
+- Updated Image URL to data links (#122)
 
 ## 5.1.0 (2024-06-23)