+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added poster for video (#120)
 - Updated current index on data series decrease (#121)
 - Updated Image URL to data links (#122)
+- Added slideshow between mixed media (#123)
 
 ## 5.1.0 (2024-06-23)